+++ MaldaLang.Tests/InterpretTranspilePairTests.cs
@@ -648,6 +648,24 @@ public class InterpretTranspilePairTests
     }
 
     [Fact]
+    public void ClassAugmentation_SameStdout()
+    {
+        InterpretTranspilePair.AssertSameFromSource(
+            """
+            class Point(x, y);
+            class Point {
+                function total() {
+                    return this.x + this.y;
+                }
+            }
+            var p = new Point(3, 4);
+            io.print(p.x);
+            io.print(p.total());
+            """,
+            "class-augmentation");
+    }
+
+    [Fact]
     public void TaggedCatch_SameStdout()
     {
         InterpretTranspilePair.AssertSameFromSource(